refactor: decompose oversized page components (#171)
Extract focused sub-components from 5 page files that exceeded 1000
lines. Each settings page becomes a thin shell delegating to tab
components; listing pages extract dialogs, filter bars, and view
components. No behavioral changes.
