+++ frontend/src/pages/ProjectSettingsPage.tsx
@@ -1,6 +1,4 @@
-import { useEffect, useMemo, useState } from "react";
-import { Link, useParams, useRouter } from "@tanstack/react-router";
-import { ColumnDef } from "@tanstack/react-table";
+import { Link, useParams } from "@tanstack/react-router";
 import { useTranslation } from "react-i18next";
 import {
   Breadcrumb,
@@ -11,465 +9,29 @@ import {
   BreadcrumbSeparator,
 } from "@/components/ui/breadcrumb";
 import { Button } from "@/components/ui/button";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardFooter,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { ConfirmDialog } from "@/components/ui/confirm-dialog";
-import { DataTable } from "@/components/ui/data-table";
-import { Label } from "@/components/ui/label";
+import { Card, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from "@/components/ui/select";
-import { Textarea } from "@/components/ui/textarea";
-import { Input } from "@/components/ui/input";
-import { SearchableCombobox } from "@/components/ui/searchable-combobox";
-import { EmojiPicker } from "@/components/EmojiPicker";
 import { useAuth } from "@/hooks/useAuth";
 import { useInitiatives } from "@/hooks/useInitiatives";
-import { useInitiativeRoles } from "@/hooks/useInitiativeRoles";
-import {
-  useProject,
-  useUpdateProject,
-  useDeleteProject,
-  useArchiveProject,
-  useUnarchiveProject,
-  useDuplicateProject,
-  useAddProjectMember,
-  useUpdateProjectMember,
-  useRemoveProjectMember,
-  useAddProjectMembersBulk,
-  useRemoveProjectMembersBulk,
-  useAddProjectRolePermission,
-  useUpdateProjectRolePermission,
-  useRemoveProjectRolePermission,
-} from "@/hooks/useProjects";
+import { useProject } from "@/hooks/useProjects";
 import { useGuildPath } from "@/lib/guildUrl";
-import {
-  ProjectPermissionLevel,
-  ProjectRolePermissionRead,
-  TagSummary,
-} from "@/api/generated/initiativeAPI.schemas";
 import { ProjectTaskStatusesManager } from "@/components/projects/ProjectTaskStatusesManager";
-import { TagPicker } from "@/components/tags";
-import { useSetProjectTags } from "@/hooks/useTags";
-
-interface PermissionRow {
-  userId: number;
-  displayName: string;
-  email: string;
-  level: ProjectPermissionLevel;
-  isOwner: boolean;
-}
+import { ProjectSettingsDetailsTab } from "@/components/projects/settings/ProjectSettingsDetailsTab";
+import { ProjectSettingsAccessTab } from "@/components/projects/settings/ProjectSettingsAccessTab";
+import { ProjectSettingsAdvancedTab } from "@/components/projects/settings/ProjectSettingsAdvancedTab";
 
 export const ProjectSettingsPage = () => {
   const { projectId } = useParams({ strict: false }) as { projectId: string };
   const parsedProjectId = Number(projectId);
-  const router = useRouter();
   const { user } = useAuth();
   const gp = useGuildPath();
   const { t } = useTranslation("projects");
-  const [selectedInitiativeId, setSelectedInitiativeId] = useState<string>("");
-  const [initiativeMessage, setInitiativeMessage] = useState<string | null>(null);
-  const [nameText, setNameText] = useState<string>("");
-  const [iconText, setIconText] = useState<string>("");
-  const [identityMessage, setIdentityMessage] = useState<string | null>(null);
-  const [descriptionText, setDescriptionText] = useState<string>("");
-  const [descriptionMessage, setDescriptionMessage] = useState<string | null>(null);
-  const [templateMessage, setTemplateMessage] = useState<string | null>(null);
-  const [duplicateMessage, setDuplicateMessage] = useState<string | null>(null);
-  const [accessMessage, setAccessMessage] = useState<string | null>(null);
-  const [accessError, setAccessError] = useState<string | null>(null);
-  const [selectedNewUserId, setSelectedNewUserId] = useState<string>("");
-  const [selectedNewLevel, setSelectedNewLevel] = useState<ProjectPermissionLevel>("read");
-  const [showDeleteConfirm, setShowDeleteConfirm] = useState(false);
-  const [selectedMembers, setSelectedMembers] = useState<PermissionRow[]>([]);
-  const [projectTags, setProjectTags] = useState<TagSummary[]>([]);
-  const [roleAccessMessage, setRoleAccessMessage] = useState<string | null>(null);
-  const [roleAccessError, setRoleAccessError] = useState<string | null>(null);
-  const [selectedNewRoleId, setSelectedNewRoleId] = useState<string>("");
-  const [selectedNewRoleLevel, setSelectedNewRoleLevel] = useState<"read" | "write">("read");
-
-  const setProjectTagsMutation = useSetProjectTags();
 
   const projectQuery = useProject(Number.isFinite(parsedProjectId) ? parsedProjectId : null);
 
   const initiativesQuery = useInitiatives({ enabled: user?.role === "admin" });
 
-  useEffect(() => {
-    if (projectQuery.data) {
-      setSelectedInitiativeId(String(projectQuery.data.initiative_id));
-      setNameText(projectQuery.data.name);
-      setIconText(projectQuery.data.icon ?? "");
-      setDescriptionText(projectQuery.data.description ?? "");
-      setProjectTags(projectQuery.data.tags ?? []);
-      setAccessMessage(null);
-      setAccessError(null);
-      setRoleAccessMessage(null);
-      setRoleAccessError(null);
-      setInitiativeMessage(null);
-      setIdentityMessage(null);
-      setDescriptionMessage(null);
-      setTemplateMessage(null);
-    }
-  }, [projectQuery.data]);
-
-  const updateInitiativeOwnership = useUpdateProject({
-    onSuccess: (data) => {
-      setInitiativeMessage(t("settings.initiative.updated"));
-      setSelectedInitiativeId(String(data.initiative_id));
-    },
-  });
-
-  const updateIdentity = useUpdateProject({
-    onSuccess: (data) => {
-      setIdentityMessage(t("settings.details.detailsUpdated"));
-      setNameText(data.name);
-      setIconText(data.icon ?? "");
-    },
-  });
-
-  const archiveProject = useArchiveProject();
-
-  const unarchiveProject = useUnarchiveProject();
-
-  const updateDescription = useUpdateProject({
-    onSuccess: (data) => {
-      setDescriptionMessage(t("settings.details.descriptionUpdated"));
-      setDescriptionText(data.description ?? "");
-    },
-  });
-
-  const duplicateProject = useDuplicateProject({
-    onSuccess: (data) => {
-      setDuplicateMessage(t("settings.duplicate.duplicated"));
-      router.navigate({ to: gp(`/projects/${data.id}`) });
-    },
-  });
-
-  const toggleTemplateStatus = useUpdateProject({
-    onSuccess: (_data, vars) => {
-      setTemplateMessage(
-        vars.data.is_template
-          ? t("settings.templateStatus.markedAsTemplate")
-          : t("settings.templateStatus.removedFromTemplates")
-      );
-    },
-  });
-
-  const deleteProject = useDeleteProject({
-    onSuccess: () => {
-      router.navigate({ to: "/" });
-    },
-  });
-
-  const addMember = useAddProjectMember(parsedProjectId, {
-    onSuccess: () => {
-      setAccessMessage(t("settings.access.granted"));
-      setAccessError(null);
-      setSelectedNewUserId("");
-      setSelectedNewLevel("read");
-    },
-    onError: () => {
-      setAccessMessage(null);
-      setAccessError(t("settings.access.grantError"));
-    },
-  });
-
-  const updateMemberLevel = useUpdateProjectMember(parsedProjectId, {
-    onSuccess: () => {
-      setAccessMessage(t("settings.access.updated"));
-      setAccessError(null);
-    },
-    onError: () => {
-      setAccessMessage(null);
-      setAccessError(t("settings.access.updateError"));
-    },
-  });
-
-  const removeMember = useRemoveProjectMember(parsedProjectId, {
-    onSuccess: () => {
-      setAccessMessage(t("settings.access.removed"));
-      setAccessError(null);
-    },
-    onError: () => {
-      setAccessMessage(null);
-      setAccessError(t("settings.access.removeError"));
-    },
-  });
-
-  const addAllMembers = useAddProjectMembersBulk(parsedProjectId, {
-    onSuccess: () => {
-      setAccessMessage(t("settings.access.grantedAll"));
-      setAccessError(null);
-      setSelectedNewUserId("");
-      setSelectedNewLevel("read");
-    },
-    onError: () => {
-      setAccessMessage(null);
-      setAccessError(t("settings.access.grantAllError"));
-    },
-  });
-
-  const bulkUpdateLevel = useAddProjectMembersBulk(parsedProjectId, {
-    onSuccess: () => {
-      setAccessMessage(t("settings.access.bulkUpdated"));
-      setAccessError(null);
-      setSelectedMembers([]);
-    },
-    onError: () => {
-      setAccessMessage(null);
-      setAccessError(t("settings.access.bulkUpdateError"));
-    },
-  });
-
-  const bulkRemoveMembers = useRemoveProjectMembersBulk(parsedProjectId, {
-    onSuccess: () => {
-      setAccessMessage(t("settings.access.bulkRemoved"));
-      setAccessError(null);
-      setSelectedMembers([]);
-    },
-    onError: () => {
-      setAccessMessage(null);
-      setAccessError(t("settings.access.bulkRemoveError"));
-    },
-  });
-
   const project = projectQuery.data;
-
-  const initiativeRolesQuery = useInitiativeRoles(project?.initiative_id ?? null);
-
-  const addRolePermission = useAddProjectRolePermission(parsedProjectId, {
-    onSuccess: () => {
-      setRoleAccessMessage(t("settings.roleAccess.granted"));
-      setRoleAccessError(null);
-      setSelectedNewRoleId("");
-      setSelectedNewRoleLevel("read");
-    },
-    onError: () => {
-      setRoleAccessMessage(null);
-      setRoleAccessError(t("settings.roleAccess.grantError"));
-    },
-  });
-
-  const updateRolePermission = useUpdateProjectRolePermission(parsedProjectId, {
-    onSuccess: () => {
-      setRoleAccessMessage(t("settings.roleAccess.updated"));
-      setRoleAccessError(null);
-    },
-    onError: () => {
-      setRoleAccessMessage(null);
-      setRoleAccessError(t("settings.roleAccess.updateError"));
-    },
-  });
-
-  const removeRolePermission = useRemoveProjectRolePermission(parsedProjectId, {
-    onSuccess: () => {
-      setRoleAccessMessage(t("settings.roleAccess.removed"));
-      setRoleAccessError(null);
-    },
-    onError: () => {
-      setRoleAccessMessage(null);
-      setRoleAccessError(t("settings.roleAccess.removeError"));
-    },
-  });
-
-  // Initiative roles not yet assigned to this project
-  const availableRoles = useMemo(
-    () =>
-      (initiativeRolesQuery.data ?? []).filter(
-        (role) => !(project?.role_permissions ?? []).some((rp) => rp.initiative_role_id === role.id)
-      ),
-    [initiativeRolesQuery.data, project?.role_permissions]
-  );
-
-  // Column definitions for role permissions table
-  const rolePermissionColumns: ColumnDef<ProjectRolePermissionRead>[] = useMemo(
-    () => [
-      {
-        accessorKey: "role_display_name",
-        header: t("settings.roleAccess.roleNameColumn"),
-        cell: ({ row }) => <span className="font-medium">{row.original.role_display_name}</span>,
-      },
-      {
-        accessorKey: "level",
-        header: t("settings.roleAccess.accessLevelColumn"),
-        cell: ({ row }) => (
-          <Select
-            value={row.original.level}
-            onValueChange={(value) => {
-              setRoleAccessMessage(null);
-              setRoleAccessError(null);
-              updateRolePermission.mutate({
-                roleId: row.original.initiative_role_id,
-                data: { level: value as ProjectPermissionLevel },
-              });
-            }}
-            disabled={updateRolePermission.isPending}
-          >
-            <SelectTrigger className="w-[130px]">
-              <SelectValue />
-            </SelectTrigger>
-            <SelectContent>
-              <SelectItem value="read">{t("settings.roleAccess.canView")}</SelectItem>
-              <SelectItem value="write">{t("settings.roleAccess.canEdit")}</SelectItem>
-            </SelectContent>
-          </Select>
-        ),
-      },
-      {
-        id: "actions",
-        header: () => <div className="text-right">{t("settings.roleAccess.actionsColumn")}</div>,
-        cell: ({ row }) => (
-          <div className="text-right">
-            <Button
-              type="button"
-              variant="ghost"
-              size="sm"
-              className="text-destructive"
-              onClick={() => {
-                setRoleAccessMessage(null);
-                setRoleAccessError(null);
-                removeRolePermission.mutate(row.original.initiative_role_id);
-              }}
-              disabled={removeRolePermission.isPending}
-            >
-              {t("settings.roleAccess.remove")}
-            </Button>
-          </div>
-        ),
-      },
-    ],
-    [updateRolePermission, removeRolePermission, t]
-  );
-
-  const initiativeMembers = useMemo(
-    () => project?.initiative?.members ?? [],
-    [project?.initiative?.members]
-  );
-
-  // Build permission rows with user info
-  const permissionRows: PermissionRow[] = useMemo(
-    () =>
-      (project?.permissions ?? []).map((permission) => {
-        const member = initiativeMembers.find((entry) => entry.user?.id === permission.user_id);
-        const ownerInfo = project?.owner;
-        const displayName =
-          member?.user?.full_name?.trim() ||
-          member?.user?.email ||
-          (permission.user_id === project?.owner_id
-            ? ownerInfo?.full_name?.trim() || ownerInfo?.email || "Project owner"
-            : `User ${permission.user_id}`);
-        const email =
-          member?.user?.email ||
-          (permission.user_id === project?.owner_id ? ownerInfo?.email || "" : "");
-        return {
-          userId: permission.user_id,
-          displayName,
-          email,
-          level: permission.level,
-          isOwner: permission.user_id === project?.owner_id,
-        };
-      }),
-    [project?.permissions, project?.owner, project?.owner_id, initiativeMembers]
-  );
-
-  // Column definitions for the permissions table
-  const permissionColumns: ColumnDef<PermissionRow>[] = useMemo(
-    () => [
-      {
-        accessorKey: "displayName",
-        header: t("settings.access.nameColumn"),
-        cell: ({ row }) => <span className="font-medium">{row.original.displayName}</span>,
-      },
-      {
-        accessorKey: "email",
-        header: t("settings.access.emailColumn"),
-        cell: ({ row }) => <span className="text-muted-foreground">{row.original.email}</span>,
-      },
-      {
-        accessorKey: "level",
-        header: t("settings.access.accessColumn"),
-        cell: ({ row }) => {
-          if (row.original.isOwner) {
-            return (
-              <span className="text-muted-foreground">{t("settings.access.permissionOwner")}</span>
-            );
-          }
-          return (
-            <Select
-              value={row.original.level}
-              onValueChange={(value) => {
-                setAccessMessage(null);
-                setAccessError(null);
-                updateMemberLevel.mutate({
-                  userId: row.original.userId,
-                  data: { level: value as ProjectPermissionLevel },
-                });
-              }}
-              disabled={updateMemberLevel.isPending}
-            >
-              <SelectTrigger className="w-[130px]">
-                <SelectValue />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="read">{t("settings.access.permissionRead")}</SelectItem>
-                <SelectItem value="write">{t("settings.access.permissionWrite")}</SelectItem>
-              </SelectContent>
-            </Select>
-          );
-        },
-      },
-      {
-        id: "actions",
-        header: () => <div className="text-right">{t("settings.access.actionsColumn")}</div>,
-        cell: ({ row }) => {
-          if (row.original.isOwner) {
-            return <div className="text-muted-foreground text-right text-xs">-</div>;
-          }
-          return (
-            <div className="text-right">
-              <Button
-                type="button"
-                variant="ghost"
-                size="sm"
-                className="text-destructive"
-                onClick={() => {
-                  setAccessMessage(null);
-                  setAccessError(null);
-                  removeMember.mutate(row.original.userId);
-                }}
-                disabled={removeMember.isPending}
-              >
-                {t("settings.access.remove")}
-              </Button>
-            </div>
-          );
-        },
-      },
-    ],
-    [updateMemberLevel, removeMember, t]
-  );
-
-  // Initiative members who don't have permissions yet
-  const availableMembers = useMemo(
-    () =>
-      initiativeMembers.filter(
-        (member) =>
-          member.user &&
-          !(project?.permissions ?? []).some((permission) => permission.user_id === member.user.id)
-      ),
-    [initiativeMembers, project?.permissions]
-  );
 
   if (!Number.isFinite(parsedProjectId)) {
     return <p className="text-destructive">{t("detail.invalidProjectId")}</p>;
@@ -564,456 +126,18 @@ export const ProjectSettingsPage = () => {
         </TabsList>
 
         {/* ── Details tab ── */}
-        <TabsContent value="details" className="space-y-6">
-          <Card className="shadow-sm">
-            <CardHeader>
-              <CardTitle>{t("settings.details.title")}</CardTitle>
-              <CardDescription>{t("settings.details.description")}</CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-8">
-              <div className="space-y-3">
-                <div className="space-y-1">
-                  <h3 className="text-base font-medium">{t("settings.details.identityHeading")}</h3>
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.details.identityDescription")}
-                  </p>
-                </div>
-                {canWriteProject ? (
-                  <form
-                    className="space-y-4"
-                    onSubmit={(event) => {
-                      event.preventDefault();
-                      setIdentityMessage(null);
-                      const trimmedIcon = iconText.trim();
-                      updateIdentity.mutate({
-                        projectId: parsedProjectId,
-                        data: {
-                          name: nameText.trim() || projectQuery.data?.name || "",
-                          icon: trimmedIcon || null,
-                        },
-                      });
-                    }}
-                  >
-                    <div className="flex flex-col gap-4 md:flex-row md:items-start">
-                      <div className="w-full space-y-2 md:max-w-xs">
-                        <Label htmlFor="project-icon">{t("settings.details.iconLabel")}</Label>
-                        <EmojiPicker
-                          id="project-icon"
-                          value={iconText || undefined}
-                          onChange={(emoji) => setIconText(emoji ?? "")}
-                        />
-                        <p className="text-muted-foreground text-sm">
-                          {t("settings.details.iconHint")}
-                        </p>
-                      </div>
-                      <div className="w-full flex-1 space-y-2">
-                        <Label htmlFor="project-name">{t("settings.details.nameLabel")}</Label>
-                        <Input
-                          id="project-name"
-                          value={nameText}
-                          onChange={(event) => setNameText(event.target.value)}
-                          placeholder={t("settings.details.namePlaceholder")}
-                          required
-                        />
-                      </div>
-                    </div>
-                    <div className="flex flex-col gap-2">
-                      <Button type="submit" disabled={updateIdentity.isPending}>
-                        {updateIdentity.isPending
-                          ? t("settings.details.saving")
-                          : t("settings.details.saveDetails")}
-                      </Button>
-                      {identityMessage ? (
-                        <p className="text-primary text-sm">{identityMessage}</p>
-                      ) : null}
-                      {updateIdentity.isError ? (
-                        <p className="text-destructive text-sm">
-                          {t("settings.details.updateError")}
-                        </p>
-                      ) : null}
-                    </div>
-                  </form>
-                ) : (
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.details.noWriteAccessIdentity")}
-                  </p>
-                )}
-              </div>
-
-              <div className="bg-border h-px" />
-
-              <div className="space-y-3">
-                <div className="space-y-1">
-                  <h3 className="text-base font-medium">
-                    {t("settings.details.descriptionHeading")}
-                  </h3>
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.details.descriptionDescription")}
-                  </p>
-                </div>
-                {canWriteProject ? (
-                  <form
-                    className="space-y-4"
-                    onSubmit={(event) => {
-                      event.preventDefault();
-                      updateDescription.mutate({
-                        projectId: parsedProjectId,
-                        data: { description: descriptionText },
-                      });
-                    }}
-                  >
-                    <Textarea
-                      rows={4}
-                      value={descriptionText}
-                      onChange={(event) => setDescriptionText(event.target.value)}
-                      placeholder={t("settings.details.descriptionPlaceholder")}
-                    />
-                    <div className="flex flex-col gap-2">
-                      <Button type="submit" disabled={updateDescription.isPending}>
-                        {updateDescription.isPending
-                          ? t("settings.details.saving")
-                          : t("settings.details.saveDescription")}
-                      </Button>
-                      {descriptionMessage ? (
-                        <p className="text-primary text-sm">{descriptionMessage}</p>
-                      ) : null}
-                    </div>
-                  </form>
-                ) : (
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.details.noWriteAccessDescription")}
-                  </p>
-                )}
-              </div>
-
-              <div className="bg-border h-px" />
-
-              <div className="space-y-3">
-                <div className="space-y-1">
-                  <h3 className="text-base font-medium">{t("settings.details.tagsHeading")}</h3>
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.details.tagsDescription")}
-                  </p>
-                </div>
-                {canWriteProject ? (
-                  <TagPicker
-                    selectedTags={projectTags}
-                    onChange={(newTags) => {
-                      setProjectTags(newTags);
-                      setProjectTagsMutation.mutate({
-                        projectId: parsedProjectId,
-                        tagIds: newTags.map((tag) => tag.id),
-                      });
-                    }}
-                  />
-                ) : (
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.details.noWriteAccessTags")}
-                  </p>
-                )}
-              </div>
-            </CardContent>
-          </Card>
-
-          {user?.role === "admin" ? (
-            <Card className="shadow-sm">
-              <CardHeader>
-                <CardTitle>{t("settings.initiative.title")}</CardTitle>
-                <CardDescription>{t("settings.initiative.description")}</CardDescription>
-              </CardHeader>
-              <CardContent>
-                {initiativesQuery.isError ? (
-                  <p className="text-destructive text-sm">{t("settings.initiative.loadError")}</p>
-                ) : (
-                  <form
-                    className="flex flex-wrap items-end gap-3"
-                    onSubmit={(event) => {
-                      event.preventDefault();
-                      if (!selectedInitiativeId) return;
-                      updateInitiativeOwnership.mutate({
-                        projectId: parsedProjectId,
-                        data: { initiative_id: Number(selectedInitiativeId) },
-                      });
-                    }}
-                  >
-                    <div className="min-w-[220px] flex-1">
-                      <Label htmlFor="project-initiative">
-                        {t("settings.initiative.owningInitiative")}
-                      </Label>
-                      <Select value={selectedInitiativeId} onValueChange={setSelectedInitiativeId}>
-                        <SelectTrigger id="project-initiative" className="mt-2">
-                          <SelectValue placeholder={t("settings.initiative.selectInitiative")} />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {initiativesQuery.data?.map((initiative) => (
-                            <SelectItem key={initiative.id} value={String(initiative.id)}>
-                              {initiative.name}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                    </div>
-                    <div className="flex flex-col gap-2">
-                      <Button type="submit" disabled={updateInitiativeOwnership.isPending}>
-                        {updateInitiativeOwnership.isPending
-                          ? t("settings.initiative.saving")
-                          : t("settings.initiative.saveInitiative")}
-                      </Button>
-                      {initiativeMessage ? (
-                        <p className="text-primary text-sm">{initiativeMessage}</p>
-                      ) : null}
-                    </div>
-                  </form>
-                )}
-              </CardContent>
-            </Card>
-          ) : null}
-        </TabsContent>
+        <ProjectSettingsDetailsTab
+          project={project}
+          projectId={parsedProjectId}
+          canWriteProject={canWriteProject}
+          isAdmin={user?.role === "admin"}
+          initiatives={initiativesQuery.data}
+          initiativesError={initiativesQuery.isError}
+        />
 
         {/* ── Access tab ── */}
         {canManageAccess ? (
-          <TabsContent value="access" className="space-y-6">
-            <Card className="shadow-sm">
-              <CardHeader>
-                <CardTitle>{t("settings.roleAccess.title")}</CardTitle>
-                <CardDescription>{t("settings.roleAccess.description")}</CardDescription>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                {(project.role_permissions ?? []).length > 0 ? (
-                  <DataTable
-                    columns={rolePermissionColumns}
-                    data={project.role_permissions ?? []}
-                    getRowId={(row) => String(row.initiative_role_id)}
-                  />
-                ) : (
-                  <p className="text-muted-foreground text-sm">
-                    {t("settings.roleAccess.noRoles")}
-                  </p>
-                )}
-
-                <div className="space-y-2 pt-2">
-                  <Label>{t("settings.roleAccess.addRole")}</Label>
-                  {initiativeRolesQuery.isLoading ? (
-                    <p className="text-muted-foreground text-sm">
-                      {t("settings.roleAccess.loadingRoles")}
-                    </p>
-                  ) : availableRoles.length === 0 ? (
-                    <p className="text-muted-foreground text-sm">
-                      {t("settings.roleAccess.allRolesAssigned")}
-                    </p>
-                  ) : (
-                    <form
-                      className="flex flex-wrap items-end gap-3"
-                      onSubmit={(event) => {
-                        event.preventDefault();
-                        if (!selectedNewRoleId) {
-                          setRoleAccessError(t("settings.roleAccess.selectRoleError"));
-                          return;
-                        }
-                        setRoleAccessError(null);
-                        addRolePermission.mutate({
-                          initiative_role_id: Number(selectedNewRoleId),
-                          level: selectedNewRoleLevel,
-                        });
-                      }}
-                    >
-                      <Select value={selectedNewRoleId} onValueChange={setSelectedNewRoleId}>
-                        <SelectTrigger className="min-w-[200px]">
-                          <SelectValue placeholder={t("settings.roleAccess.selectRole")} />
-                        </SelectTrigger>
-                        <SelectContent>
-                          {availableRoles.map((role) => (
-                            <SelectItem key={role.id} value={String(role.id)}>
-                              {role.display_name}
-                            </SelectItem>
-                          ))}
-                        </SelectContent>
-                      </Select>
-                      <Select
-                        value={selectedNewRoleLevel}
-                        onValueChange={(value) =>
-                          setSelectedNewRoleLevel(value as "read" | "write")
-                        }
-                      >
-                        <SelectTrigger className="w-[130px]">
-                          <SelectValue />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="read">{t("settings.roleAccess.canView")}</SelectItem>
-                          <SelectItem value="write">{t("settings.roleAccess.canEdit")}</SelectItem>
-                        </SelectContent>
-                      </Select>
-                      <Button type="submit" disabled={addRolePermission.isPending}>
-                        {addRolePermission.isPending
-                          ? t("settings.roleAccess.adding")
-                          : t("settings.roleAccess.add")}
-                      </Button>
-                    </form>
-                  )}
-                  {roleAccessMessage ? (
-                    <p className="text-primary text-sm">{roleAccessMessage}</p>
-                  ) : null}
-                  {roleAccessError ? (
-                    <p className="text-destructive text-sm">{roleAccessError}</p>
-                  ) : null}
-                </div>
-              </CardContent>
-            </Card>
-
-            <Card className="shadow-sm">
-              <CardHeader>
-                <CardTitle>{t("settings.access.title")}</CardTitle>
-                <CardDescription>{t("settings.access.description")}</CardDescription>
-              </CardHeader>
-              <CardContent className="space-y-4">
-                {/* Bulk action bar */}
-                {selectedMembers.length > 0 && (
-                  <div className="bg-muted flex items-center gap-3 rounded-md p-3">
-                    <span className="text-sm font-medium">
-                      {t("settings.access.selected", { count: selectedMembers.length })}
-                    </span>
-                    <Select
-                      onValueChange={(level) => {
-                        const userIds = selectedMembers
-                          .filter((m) => !m.isOwner)
-                          .map((m) => m.userId);
-                        if (userIds.length > 0) {
-                          bulkUpdateLevel.mutate({
-                            user_ids: userIds,
-                            level: level as ProjectPermissionLevel,
-                          });
-                        }
-                      }}
-                      disabled={bulkUpdateLevel.isPending || bulkRemoveMembers.isPending}
-                    >
-                      <SelectTrigger className="w-[150px]">
-                        <SelectValue placeholder={t("settings.access.changeAccess")} />
-                      </SelectTrigger>
-                      <SelectContent>
-                        <SelectItem value="read">{t("settings.access.permissionRead")}</SelectItem>
-                        <SelectItem value="write">
-                          {t("settings.access.permissionWrite")}
-                        </SelectItem>
-                      </SelectContent>
-                    </Select>
-                    <Button
-                      type="button"
-                      variant="destructive"
-                      size="sm"
-                      onClick={() => {
-                        const userIds = selectedMembers
-                          .filter((m) => !m.isOwner)
-                          .map((m) => m.userId);
-                        if (userIds.length > 0) {
-                          bulkRemoveMembers.mutate({ user_ids: userIds });
-                        }
-                      }}
-                      disabled={bulkUpdateLevel.isPending || bulkRemoveMembers.isPending}
-                    >
-                      {bulkRemoveMembers.isPending
-                        ? t("settings.access.removing")
-                        : t("settings.access.remove")}
-                    </Button>
-                  </div>
-                )}
-
-                {/* Access table */}
-                <DataTable
-                  columns={permissionColumns}
-                  data={permissionRows}
-                  enablePagination
-                  enableFilterInput
-                  filterInputColumnKey="displayName"
-                  filterInputPlaceholder={t("settings.access.filterByName")}
-                  enableRowSelection
-                  onRowSelectionChange={setSelectedMembers}
-                  onExitSelection={() => setSelectedMembers([])}
-                  getRowId={(row) => String(row.userId)}
-                />
-
-                {/* Add member form */}
-                <div className="space-y-2 pt-2">
-                  <Label>{t("settings.access.grantAccess")}</Label>
-                  {availableMembers.length === 0 ? (
-                    <p className="text-muted-foreground text-sm">
-                      {t("settings.access.allMembersHaveAccess")}
-                    </p>
-                  ) : (
-                    <form
-                      className="flex flex-wrap items-end gap-3"
-                      onSubmit={(event) => {
-                        event.preventDefault();
-                        if (!selectedNewUserId) {
-                          setAccessError(t("settings.access.selectMemberError"));
-                          return;
-                        }
-                        setAccessError(null);
-                        addMember.mutate({
-                          user_id: Number(selectedNewUserId),
-                          level: selectedNewLevel,
-                        });
-                      }}
-                    >
-                      <SearchableCombobox
-                        items={availableMembers.map((member) => ({
-                          value: String(member.user.id),
-                          label: member.user.full_name?.trim() || member.user.email,
-                        }))}
-                        value={selectedNewUserId}
-                        onValueChange={setSelectedNewUserId}
-                        placeholder={t("settings.access.selectMember")}
-                        emptyMessage={t("settings.access.noMembersFound")}
-                        className="min-w-[200px]"
-                      />
-                      <Select
-                        value={selectedNewLevel}
-                        onValueChange={(value) =>
-                          setSelectedNewLevel(value as ProjectPermissionLevel)
-                        }
-                      >
-                        <SelectTrigger className="w-[130px]">
-                          <SelectValue />
-                        </SelectTrigger>
-                        <SelectContent>
-                          <SelectItem value="read">
-                            {t("settings.access.permissionRead")}
-                          </SelectItem>
-                          <SelectItem value="write">
-                            {t("settings.access.permissionWrite")}
-                          </SelectItem>
-                        </SelectContent>
-                      </Select>
-                      <Button
-                        type="submit"
-                        disabled={addMember.isPending || addAllMembers.isPending}
-                      >
-                        {addMember.isPending
-                          ? t("settings.access.adding")
-                          : t("settings.access.add")}
-                      </Button>
-                      <Button
-                        type="button"
-                        variant="secondary"
-                        onClick={() =>
-                          addAllMembers.mutate({
-                            user_ids: availableMembers.map((m) => m.user.id),
-                            level: selectedNewLevel,
-                          })
-                        }
-                        disabled={addMember.isPending || addAllMembers.isPending}
-                      >
-                        {addAllMembers.isPending
-                          ? t("settings.access.addingAll")
-                          : t("settings.access.addAll", { count: availableMembers.length })}
-                      </Button>
-                    </form>
-                  )}
-                  {accessMessage ? <p className="text-primary text-sm">{accessMessage}</p> : null}
-                  {accessError ? <p className="text-destructive text-sm">{accessError}</p> : null}
-                </div>
-              </CardContent>
-            </Card>
-          </TabsContent>
+          <ProjectSettingsAccessTab project={project} projectId={parsedProjectId} />
         ) : null}
 
         {/* ── Task statuses tab ── */}
@@ -1025,158 +149,13 @@ export const ProjectSettingsPage = () => {
         </TabsContent>
 
         {/* ── Advanced tab ── */}
-        <TabsContent value="advanced" className="space-y-6">
-          <Card className="shadow-sm">
-            <CardHeader>
-              <CardTitle>{t("settings.templateStatus.title")}</CardTitle>
-              <CardDescription>{t("settings.templateStatus.description")}</CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-2">
-              <p className="text-muted-foreground text-sm">
-                {project.is_template
-                  ? t("settings.templateStatus.isTemplate")
-                  : t("settings.templateStatus.isStandard")}
-              </p>
-              {templateMessage ? <p className="text-primary text-sm">{templateMessage}</p> : null}
-            </CardContent>
-            <CardFooter className="flex flex-wrap gap-3">
-              {canWriteProject ? (
-                <Button
-                  type="button"
-                  variant={project.is_template ? "outline" : "default"}
-                  onClick={() => {
-                    setTemplateMessage(null);
-                    toggleTemplateStatus.mutate({
-                      projectId: parsedProjectId,
-                      data: { is_template: !project.is_template },
-                    });
-                  }}
-                  disabled={toggleTemplateStatus.isPending}
-                >
-                  {project.is_template
-                    ? t("settings.templateStatus.convertToStandard")
-                    : t("settings.templateStatus.markAsTemplate")}
-                </Button>
-              ) : (
-                <p className="text-muted-foreground text-sm">
-                  {t("settings.templateStatus.noWriteAccess")}
-                </p>
-              )}
-              {project.is_template ? (
-                <Button asChild variant="link" className="px-0">
-                  <Link to={gp("/projects")}>{t("settings.templateStatus.viewAllTemplates")}</Link>
-                </Button>
-              ) : null}
-            </CardFooter>
-          </Card>
-
-          <Card className="shadow-sm">
-            <CardHeader>
-              <CardTitle>{t("settings.duplicate.title")}</CardTitle>
-              <CardDescription>{t("settings.duplicate.description")}</CardDescription>
-            </CardHeader>
-            <CardContent>
-              {duplicateMessage ? <p className="text-primary text-sm">{duplicateMessage}</p> : null}
-            </CardContent>
-            <CardFooter>
-              {canWriteProject ? (
-                <Button
-                  type="button"
-                  onClick={() => {
-                    const defaultName = `${project.name} copy`;
-                    const newName = window.prompt(t("settings.duplicate.promptName"), defaultName);
-                    if (newName === null) {
-                      return;
-                    }
-                    setDuplicateMessage(null);
-                    duplicateProject.mutate({
-                      projectId: parsedProjectId,
-                      data: { name: newName.trim() || undefined },
-                    });
-                  }}
-                  disabled={duplicateProject.isPending}
-                >
-                  {duplicateProject.isPending
-                    ? t("settings.duplicate.duplicating")
-                    : t("settings.duplicate.duplicateButton")}
-                </Button>
-              ) : (
-                <p className="text-muted-foreground text-sm">
-                  {t("settings.duplicate.noWriteAccess")}
-                </p>
-              )}
-            </CardFooter>
-          </Card>
-
-          <Card className="shadow-sm">
-            <CardHeader>
-              <CardTitle>{t("settings.archiveStatus.title")}</CardTitle>
-              <CardDescription>
-                {project.is_archived
-                  ? t("settings.archiveStatus.isArchived")
-                  : t("settings.archiveStatus.isActive")}
-              </CardDescription>
-            </CardHeader>
-            <CardFooter>
-              {canWriteProject ? (
-                <Button
-                  type="button"
-                  variant="outline"
-                  onClick={() =>
-                    project.is_archived
-                      ? unarchiveProject.mutate(parsedProjectId)
-                      : archiveProject.mutate(parsedProjectId)
-                  }
-                  disabled={archiveProject.isPending || unarchiveProject.isPending}
-                >
-                  {project.is_archived
-                    ? t("settings.archiveStatus.unarchive")
-                    : t("settings.archiveStatus.archive")}
-                </Button>
-              ) : (
-                <p className="text-muted-foreground text-sm">
-                  {t("settings.archiveStatus.noWriteAccess")}
-                </p>
-              )}
-            </CardFooter>
-          </Card>
-
-          {isOwner ? (
-            <Card className="border-destructive/40 bg-destructive/5 shadow-sm">
-              <CardHeader>
-                <CardTitle className="text-destructive">{t("settings.danger.title")}</CardTitle>
-                <CardDescription className="text-destructive">
-                  {t("settings.danger.description")}
-                </CardDescription>
-              </CardHeader>
-              <CardFooter>
-                <Button
-                  type="button"
-                  variant="destructive"
-                  onClick={() => setShowDeleteConfirm(true)}
-                  disabled={deleteProject.isPending}
-                >
-                  {t("settings.danger.deleteButton")}
-                </Button>
-              </CardFooter>
-            </Card>
-          ) : null}
-        </TabsContent>
+        <ProjectSettingsAdvancedTab
+          project={project}
+          projectId={parsedProjectId}
+          canWriteProject={canWriteProject}
+          isOwner={isOwner}
+        />
       </Tabs>
-
-      <ConfirmDialog
-        open={showDeleteConfirm}
-        onOpenChange={setShowDeleteConfirm}
-        title={t("settings.danger.deleteTitle")}
-        description={t("settings.danger.deleteDescription")}
-        confirmLabel={t("settings.danger.deleteConfirm")}
-        onConfirm={() => {
-          deleteProject.mutate(parsedProjectId);
-          setShowDeleteConfirm(false);
-        }}
-        isLoading={deleteProject.isPending}
-        destructive
-      />
     </div>
   );
 };
